fix(mysql): odblokuj bramkę rc31 — trzy błędy, które SQLite przepuszcza
Wdrożenie rc31 stanęło na bramce testów: osiem porażek na MySQL przy
zielonym suicie lokalnie. Testy chodzą na SQLite, który nie egzekwuje
kluczy obcych przy usuwaniu tabel i inaczej typuje wyniki sumowania.

Testy wycofania migracji próbowały usunąć tabelę, na którą wskazuje klucz
obcy z migracji dodanej później. Prawdziwy rollback nigdy tego nie robi —
Laravel cofa partię w odwrotnej kolejności, więc dziecko znika przed
rodzicem. Błędne były testy, nie migracje: cofały pojedynczą migrację ze
środka zależności. Teraz cofają zależne w prawdziwej kolejności.

Wycofanie migracji dokładającej powiązanie egzemplarza z pozycją zamówienia
było niewykonalne na MySQL. Usunięcie indeksu przed usunięciem klucza obcego
zostawia InnoDB ograniczenie bez indeksu, bo po dołożeniu indeksu złożonego
to właśnie on spełnia wymóg klucza — osobnego już nie ma. Kolejność jest
teraz odwrotna: klucz, indeks, kolumna.

item_count w analityce zapisywał się jako tekst. Sumowanie po kolumnie
całkowitej ma w MySQL typ wynikowy DECIMAL, więc sterownik oddaje string,
podczas gdy SQLite zwraca liczbę — do kolumny JSON szło "3" zamiast 3.
Dotyczyło to UAT i produkcji, nie tylko testów. Test był poprawny od
początku i nie miał jak zapłonąć lokalnie. Przeszukane wszystkie miejsca
sumujące w aplikacji: innych podatnych nie ma.

Trzy testy w onboardingu padały wtórnie. DDL w MySQL commituje niejawnie,
więc nieudane wycofanie utrwalało wiersze utworzone wcześniej w tym samym
teście, a wycofanie transakcji nie miało już czego cofać. Naprawa poprzednich
usunęła je bez dotykania tych plików.

Zweryfikowane na jednorazowym kontenerze mysql:8.0 w obie strony: bez
poprawki osiem porażek co do znaku takich jak na bramce, z poprawką 227
zielonych. SQLite bez regresji.

+++ app/Jobs/MarkCartsAbandonedJob.php
@@ -32,7 +32,12 @@ class MarkCartsAbandonedJob implements ShouldQueue
 
                     $dispatcher->trackForCart($cart, 'cart.abandoned', [
                         'cart_id' => $cart->id,
-                        'item_count' => $cart->items_sum_quantity ?? 0,
+                        // (int) cast: MySQL's PDO driver returns SUM() over an
+                        // integer column as a numeric STRING (DECIMAL result
+                        // type), unlike SQLite — left uncast, item_count would
+                        // json_encode as a JSON string ("3") instead of a
+                        // number (3) in analytics_events.properties.
+                        'item_count' => (int) ($cart->items_sum_quantity ?? 0),
                         'checkout_started' => $cart->checkout_started_at !== null,
                         'last_step' => $cart->last_checkout_step,
                     ]);
